% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdim.R
\name{parse_epoch}
\alias{parse_epoch}
\title{Parse the epoch}
\usage{
parse_epoch(x = "days since 1800-01-01 00:00:00")
}
\arguments{
\item{x}{the "units string value form a \code{ncdim4} object}
}
\value{
list with time (POSIXct), unit, and multiplier
}
\description{
Parse the epoch
}
