% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdim.R
\name{parse_time}
\alias{parse_time}
\title{Parse the time}
\usage{
parse_time(x)
}
\arguments{
\item{x}{ncdim4 class object with name time}
}
\value{
POSIXct
}
\description{
Time if denoted in the time$units element.  It the the form
"UNITS since EPOCH"
}
