% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rarr.R
\name{Rarr}
\alias{Rarr}
\title{R6 Class Representing a NetCDF resource}
\description{
A Rarr object provides accessors for dimension and variable data.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{filename}}{chr, the filename or URL (for OPeNDAP)}

\item{\code{nc}}{ncdf4, the ncdf4 object we are navigating}

\item{\code{dims}}{named list}

\item{\code{vars}}{named list}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Rarr-new}{\code{Rarr$new()}}
\item \href{#method-Rarr-open}{\code{Rarr$open()}}
\item \href{#method-Rarr-close}{\code{Rarr$close()}}
\item \href{#method-Rarr-print}{\code{Rarr$print()}}
\item \href{#method-Rarr-dump}{\code{Rarr$dump()}}
\item \href{#method-Rarr-clone}{\code{Rarr$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rarr-new"></a>}}
\if{latex}{\out{\hypertarget{method-Rarr-new}{}}}
\subsection{Method \code{new()}}{
initialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rarr$new(filename = oisst_example_url())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{filename}}{chr, the filename or URL (for OPeNDAP)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rarr-open"></a>}}
\if{latex}{\out{\hypertarget{method-Rarr-open}{}}}
\subsection{Method \code{open()}}{
Open the NetCDF resource
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rarr$open()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rarr-close"></a>}}
\if{latex}{\out{\hypertarget{method-Rarr-close}{}}}
\subsection{Method \code{close()}}{
Close the NetCDF resource (done automatically upon gargabe cleanup)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rarr$close()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rarr-print"></a>}}
\if{latex}{\out{\hypertarget{method-Rarr-print}{}}}
\subsection{Method \code{print()}}{
print a description of the object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rarr$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{unused
Print a long summary of the NetCDF object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rarr-dump"></a>}}
\if{latex}{\out{\hypertarget{method-Rarr-dump}{}}}
\subsection{Method \code{dump()}}{
Dump a summary ala print.ncdf4
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rarr$dump()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rarr-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Rarr-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rarr$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
