% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rray-package.R
\docType{package}
\name{rarray}
\alias{rray-package}
\alias{rarray}
\alias{rray}
\title{R package to ease access to NetCDF data sources}
\description{
R package to ease access to NetCDF data sources.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/BigelowLab/rray}
}

}
