% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rdim.R
\name{Rdim}
\alias{Rdim}
\title{R6 Class Representing a NetCDF dimension attribute}
\description{
A Rdim object provides accessors for dimension data
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ncdim4}}{list}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{dimension name}

\item{\code{units}}{dimension units}

\item{\code{vals}}{dimension values (vector)}

\item{\code{len}}{length of the dimension}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Rdim-new}{\code{Rdim$new()}}
\item \href{#method-Rdim-print}{\code{Rdim$print()}}
\item \href{#method-Rdim-clone}{\code{Rdim$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rdim-new"></a>}}
\if{latex}{\out{\hypertarget{method-Rdim-new}{}}}
\subsection{Method \code{new()}}{
intialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rdim$new(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{ncdim4 object (it's a list)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rdim-print"></a>}}
\if{latex}{\out{\hypertarget{method-Rdim-print}{}}}
\subsection{Method \code{print()}}{
Print a long summary of the NetCDF dimensions
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rdim$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{unused}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rdim-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Rdim-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rdim$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
