% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rvar.R
\name{Rvar}
\alias{Rvar}
\title{R6 Class Representing a NetCDF variable attribute}
\description{
A Rdim object provides accessors for variable data
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ncvar4}}{ncvar4 class object (it's a list)}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{name}}{variable name}

\item{\code{ndims}}{number of dims}

\item{\code{size}}{dim sizes}

\item{\code{units}}{variable units}

\item{\code{longname}}{variable long name}

\item{\code{dimnames}}{dimension names}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Rvar-new}{\code{Rvar$new()}}
\item \href{#method-Rvar-print}{\code{Rvar$print()}}
\item \href{#method-Rvar-clone}{\code{Rvar$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rvar-new"></a>}}
\if{latex}{\out{\hypertarget{method-Rvar-new}{}}}
\subsection{Method \code{new()}}{
intialize
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rvar$new(x)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{x}}{ncvar4 object (it's a list)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rvar-print"></a>}}
\if{latex}{\out{\hypertarget{method-Rvar-print}{}}}
\subsection{Method \code{print()}}{
Print a long summary of the NetCDF variable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rvar$print(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{unused}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Rvar-clone"></a>}}
\if{latex}{\out{\hypertarget{method-Rvar-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Rvar$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
